+++ example/lib/src/common/constant/pubspec.yaml.g.dart
@@ -93,13 +93,13 @@ sealed class Pubspec {
   static const PubspecVersion version = (
     /// Non-canonical string representation of the version as provided
     /// in the pubspec.yaml file.
-    representation: r'0.2.7',
+    representation: r'0.2.6',
 
     /// Returns a 'canonicalized' representation
     /// of the application version.
     /// This represents the version string in accordance with
     /// Semantic Versioning (SemVer) standards.
-    canonical: r'0.2.7',
+    canonical: r'0.2.6',
 
     /// MAJOR version when you make incompatible API changes.
     /// The major version number: 1 in "1.2.3".
@@ -112,7 +112,7 @@ sealed class Pubspec {
 
     /// PATCH version when you make backward compatible bug fixes.
     /// The patch version number: 3 in "1.2.3".
-    patch: 7,
+    patch: 6,
 
     /// The pre-release identifier: "foo" in "1.2.3-foo".
     preRelease: <String>[],
@@ -127,10 +127,10 @@ sealed class Pubspec {
     5,
     6,
     12,
-    47,
-    21,
-    769,
-    647,
+    53,
+    45,
+    748,
+    101,
   );
 
   /// Name