+++ test/flutter_simple_country_picker_test.dart
@@ -5,13 +5,13 @@ import 'unit_test/countries_controller_test.dart'
 import 'unit_test/countries_parser_test.dart' as unit_countries_parser_test;
 import 'unit_test/countries_provider_test.dart' as unit_country_provider_test;
 import 'unit_test/countries_state_test.dart' as unit_countries_state_test;
-import 'unit_test/countries_util_test.dart' as unit_countries_util_test;
 import 'unit_test/country_codes_test.dart' as unit_country_codes_test;
 import 'unit_test/country_input_formater_test.dart'
     as unit_country_input_formater_test;
 import 'unit_test/country_picker_theme_test.dart'
     as unit_country_picker_theme_test;
 import 'unit_test/country_test.dart' as unit_country_test;
+import 'unit_test/country_util_test.dart' as unit_countries_util_test;
 import 'widget_test/countries_parser_test.dart' as widget_countries_parser_test;
 import 'widget_test/countries_scope_test.dart' as widget_countries_scope_test;
 import 'widget_test/country_input_formater_test.dart'